colocate transfomring, create "core" map
