+++ packages/client/package.json
@@ -59,11 +59,6 @@
       "import": "./dist/links/wsLink.mjs",
       "require": "./dist/links/wsLink.js",
       "default": "./dist/links/wsLink.js"
-    },
-    "./shared": {
-      "import": "./dist/shared/index.mjs",
-      "require": "./dist/shared/index.js",
-      "default": "./dist/shared/index.js"
     }
   },
   "files": [
@@ -72,7 +67,6 @@
     "README.md",
     "package.json",
     "links",
-    "shared",
     "!**/*.test.*"
   ],
   "dependencies": {
