+++ packages/client/rollup.config.ts
@@ -8,7 +8,6 @@ export const input = [
   'src/links/splitLink.ts',
   'src/links/loggerLink.ts',
   'src/links/wsLink.ts',
-  'src/shared/index.ts',
 ];
 
 export default function rollup(): RollupOptions[] {
