--- packages/client/src/shared/index.ts
@@ -1,1 +1,0 @@
-export * from './transformResult';
