+++ packages/client/turbo.json
@@ -2,8 +2,6 @@
   "$schema": "https://turborepo.org/schema.json",
   "extends": ["//"],
   "pipeline": {
-    "codegen-entrypoints": {
-      "outputs": ["package.json", "links/**", "shared/**"]
-    }
+    "codegen-entrypoints": { "outputs": ["package.json", "links/**"] }
   }
 }
