+++ packages/core/src/core/router.ts
@@ -1,12 +1,12 @@
-import { defaultFormatter } from './error/formatter';
-import { TRPCError } from './error/TRPCError';
-import type { ProcedureCallOptions } from './internals/procedureBuilder';
+import { createRecursiveProxy } from '../createProxy';
+import { defaultFormatter } from '../error/formatter';
+import { TRPCError } from '../error/TRPCError';
+import type { ProcedureCallOptions } from '../internals/procedureBuilder';
+import { defaultTransformer } from '../transformer';
+import { omitPrototype } from '../utilityFunctions';
+import type { MaybePromise } from '../utilityTypes';
 import type { AnyProcedure, ProcedureArgs } from './procedure';
 import type { AnyRootConfig } from './rootConfig';
-import { createRecursiveProxy } from './shared/createProxy';
-import { defaultTransformer } from './transformer';
-import { omitPrototype } from './utilityFunctions';
-import type { MaybePromise } from './utilityTypes';
 
 /** @internal **/
 export type ProcedureRecord = Record<string, AnyProcedure>;
