+++ packages/core/src/error/formatter.ts
@@ -1,4 +1,4 @@
-import type { ProcedureType } from '../procedure';
+import type { ProcedureType } from '../core/procedure';
 import type {
   TRPC_ERROR_CODE_KEY,
   TRPC_ERROR_CODE_NUMBER,
@@ -9,9 +9,10 @@ import type { TRPCError } from './TRPCError';
 /**
  * @internal
  */
-export type ErrorFormatter<TContext, TShape extends TRPCErrorShape<number>> = ({
-  error,
-}: {
+export type ErrorFormatter<
+  TContext,
+  TShape extends TRPCErrorShape<number>,
+> = (args: {
   error: TRPCError;
   type: ProcedureType | 'unknown';
   path: string | undefined;
@@ -45,10 +46,6 @@ export interface DefaultErrorShape
   code: TRPC_ERROR_CODE_NUMBER;
 }
 
-export const defaultFormatter: ErrorFormatter<any, any> = ({
-  shape,
-}: {
-  shape: DefaultErrorShape;
-}) => {
+export const defaultFormatter: ErrorFormatter<any, any> = ({ shape }) => {
   return shape;
 };
