+++ packages/core/src/http/contentType.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
+import type { AnyRouter } from '../core/router';
 import { TRPCError } from '../error/TRPCError';
-import type { AnyRouter } from '../router';
 import type { CombinedDataTransformer } from '../transformer';
 import type { MaybePromise } from '../utilityTypes';
 import type { HTTPRequest } from './types';
