+++ packages/core/src/http/resolveHTTPResponse.ts
@@ -1,11 +1,11 @@
+import type { ProcedureType } from '../core/procedure';
+import type { AnyRouter } from '../core/router';
+import { callProcedure } from '../core/router';
+import { getErrorShape } from '../error/getErrorShape';
 import { getTRPCErrorFromUnknown, TRPCError } from '../error/TRPCError';
-import type { ProcedureType } from '../procedure';
-import type { AnyRouter } from '../router';
-import { callProcedure } from '../router';
-import type { TRPCResponse } from '../rpc';
-import { getErrorShape } from '../shared/getErrorShape';
-import { transformTRPCResponse } from '../shared/transformTRPCResponse';
 import type { inferRouterContext, inferRouterError } from '../inference';
+import type { TRPCResponse } from '../rpc';
+import { transformTRPCResponse } from '../transformer';
 import type { Maybe } from '../utilityTypes';
 import type { BaseContentTypeHandler } from './contentType';
 import { getJsonContentTypeInputs } from './contentType';
