+++ packages/core/src/internals/mergeRouters.ts
@@ -1,6 +1,11 @@
+import type {
+  AnyRouter,
+  AnyRouterDef,
+  Router,
+  RouterDef,
+} from '../core/router';
+import { createRouterFactory } from '../core/router';
 import { defaultFormatter } from '../error/formatter';
-import type { AnyRouter, AnyRouterDef, Router, RouterDef } from '../router';
-import { createRouterFactory } from '../router';
 import type { CombinedDataTransformer } from '../transformer';
 import { defaultTransformer } from '../transformer';
 import { mergeWithoutOverrides } from '../utilityFunctions';
