+++ packages/core/src/serialize.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import type { Simplify, WithoutIndexSignature } from '../utilityTypes';
+import type { Simplify, WithoutIndexSignature } from './utilityTypes';
 
 /**
  * @link https://github.com/remix-run/remix/blob/2248669ed59fd716e267ea41df5d665d4781f4a9/packages/remix-server-runtime/serialize.ts
